Move SysErrorDetail and ExceptionErrorDetail classes to their own header files
