+++ src/servers/Server.cc
@@ -12,6 +12,7 @@
 #include "comm.h"
 #include "comm/Read.h"
 #include "Debug.h"
+#include "error/SysErrorDetail.h"
 #include "fd.h"
 #include "fde.h"
 #include "http/Stream.h"
