+++ src/tests/stub_liberror.cc
@@ -11,10 +11,4 @@ void Error::update(const Error &) STUB_NOP
 
 std::ostream &operator <<(std::ostream &os, const Error &) STUB_RETVAL(os)
 
-SBuf ExceptionErrorDetail::brief() const STUB_RETVAL(SBuf())
-SBuf ExceptionErrorDetail::verbose(const HttpRequestPointer &) const STUB_RETVAL(SBuf())
-
-SBuf SysErrorDetail::brief() const STUB_RETVAL(SBuf())
-SBuf SysErrorDetail::verbose(const HttpRequestPointer &) const STUB_RETVAL(SBuf())
-
 ErrorDetail::Pointer MakeNamedErrorDetail(const char *name) STUB_RETVAL(ErrorDetail::Pointer())
